a C function that prints some basic info about Python lists

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -7,7 +7,7 @@
  */
 void print_python_list_info(PyObject *p)
 {
-	long int size = PyList_Size(p);
+	int size = PyList_Size(p);
 	int alloc, i;
 	PyObject *obj;
